// manufacturingTerminal/frontendManufacturingTerminal/src/app/interfaces/enum-validation-rules.ts
export enum ValidationRule {
  Pattern = "pattern",
  Required = "required",
  Min = "min",
  Max = "max",
  MinLength = "minLength",
  MaxLength = "maxLength",
  PatternFromCamunda = "pattern",
  RequiredFromCamunda = "required",
  MinFromCamunda = "min",
  MaxFromCamunda = "max",
  MinLengthFromCamunda = "minlength",
  MaxLengthFromCamunda = "maxlength",
}